requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
backports.zoneinfo==0.2.1
boto3==1.26.109
botocore==1.29.109
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==4.1.7
django-allauth==0.39.1
django-cors-headers==3.14.0
django-crispy-forms==1.7.2
django-environ==0.10.0
Django-Gtts==0.4
django-jazzmin==2.6.0
django-phonenumber-field==7.0.2
django-rest-swagger==2.2.0
django-storages==1.13.2
djangorestframework==3.14.0
drf-spectacular==0.26.0
drf-yasg==1.21.5
et-xmlfile==1.1.0
frozenlist==1.3.3
gTTS==2.3.1
idna==3.4
importlib-resources==5.12.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.24.2
oauthlib==3.0.1
openai==0.27.0
openapi-codec==1.3.2
opencv-python==4.7.0.72
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
phonenumbers==8.13.6
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
psycopg2-binary==2.9.5
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.0
python3-openid==3.1.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.2.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
simplejson==3.18.4
six==1.16.0
SpeechRecognition==3.9.0
sqlparse==0.4.3
stripe==5.2.0
tqdm==4.65.0
uritemplate==4.1.1
urllib3==1.26.14
XlsxWriter==3.0.9
yarl==1.8.2
zipp==3.15.0
